+++ src/GraphQL/Mutations/AuthenticateModel.php
@@ -5,8 +5,8 @@ namespace Marqant\AuthGraphQL\GraphQL\Mutations;
 use Exception;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Validation\ValidationException;
-use Tjventurini\GraphQLExceptions\Exceptions\ClientSaveInternalGraphQLException;
-use Tjventurini\GraphQLExceptions\Exceptions\ClientSaveValidationGraphQLException;
+use Marqant\AuthGraphQL\Exceptions\ClientSaveInternalGraphQLException;
+use Marqant\AuthGraphQL\Exceptions\ClientSaveValidationGraphQLException;
 
 /**
  * Class AuthenticateModel
@@ -23,7 +23,7 @@ class AuthenticateModel
      *
      * @throws Exception
      */
-    public function resolve($rootValue, array $args)
+    public function __invoke($rootValue, array $args)
     {
         try {
             $table_name = $args['tableName'];